refactor: improve components structure
