--- client/src/modules/Opportunities/components/EditViewResume.tsx
@@ -1,48 +1,0 @@
-import { ResumeDto } from 'api';
-import { transformCvData } from '../transformers/transformCvData';
-import EditCV from './EditCV';
-import { NoConsentView } from './NoConsentView';
-import ViewCV from './ViewCV';
-
-type ResumeProps = {
-  githubId: string;
-  consent: boolean;
-  error?: Error;
-  data: ResumeDto | null;
-  editMode: boolean;
-  switchView: () => void;
-  onRemoveConsent: () => void;
-  onCreateConsent: () => void;
-  onUpdateResume?: () => void;
-};
-
-export function EditViewResume(props: ResumeProps) {
-  const { githubId, data, consent, editMode, switchView, onUpdateResume, onRemoveConsent, onCreateConsent } = props;
-
-  if (!consent) {
-    return <NoConsentView isOwner={true} giveConsent={onCreateConsent} />;
-  }
-
-  const { userData, contacts, visibleCourses, courses } = transformCvData(data);
-
-  const editing = editMode || data == null;
-
-  return (
-    <>
-      {editing ? (
-        <EditCV
-          courses={courses}
-          userData={userData}
-          contacts={contacts}
-          visibleCourses={visibleCourses}
-          onUpdateResume={onUpdateResume}
-          githubId={githubId}
-          onRemoveConsent={onRemoveConsent}
-          switchView={switchView}
-        />
-      ) : (
-        <ViewCV onRemoveConsent={onRemoveConsent} switchView={switchView} initialData={data} />
-      )}
-    </>
-  );
-}
