+++ client/src/modules/Opportunities/components/Link/index.tsx
@@ -1,0 +1,15 @@
+type Props = {
+  url: string;
+  title?: string;
+  text?: string;
+};
+
+export const Link = ({ url, text, title }: Props) => {
+  return url ? (
+    <a title={title} className="rs-link" target="_blank" rel="nofollow" href={url}>
+      {text}
+    </a>
+  ) : (
+    <>{text}</>
+  );
+};
