+++ client/src/modules/Opportunities/components/NoConsentView/index.test.tsx
