+++ client/src/modules/Opportunities/components/NoConsentView/index.tsx
