+++ client/src/modules/Opportunities/components/StudentStatus/index.tsx
@@ -1,0 +1,27 @@
+import { SafetyCertificateTwoTone, ClockCircleTwoTone, CheckCircleTwoTone } from '@ant-design/icons';
+
+export const StudentStatus = (props: { certificateId: string | null; isCourseCompleted: boolean }) => {
+  const { certificateId, isCourseCompleted } = props;
+
+  if (certificateId) {
+    return (
+      <span>
+        <SafetyCertificateTwoTone twoToneColor="#52c41a" /> Completed with{' '}
+        <a target="_blank" rel="nofollow" href={`/certificate/${certificateId}`}>
+          certificate
+        </a>
+      </span>
+    );
+  }
+  if (isCourseCompleted)
+    return (
+      <span>
+        <CheckCircleTwoTone /> Completed
+      </span>
+    );
+  return (
+    <span>
+      <ClockCircleTwoTone twoToneColor="#ec9607" /> In Progress
+    </span>
+  );
+};
