+++ client/src/modules/Opportunities/components/ViewCv/DataTextValue/index.tsx
@@ -1,0 +1,21 @@
+import { PropsWithChildren } from 'react';
+import css from 'styled-jsx/css';
+
+export const DataTextValue = ({ children }: PropsWithChildren<unknown>) => {
+  return (
+    <>
+      <span className="course-data-key">{children}</span>
+      <style jsx>{styles}</style>
+    </>
+  );
+};
+
+const styles = css`
+  .course-data-key {
+    font-size: 14px;
+    padding-right: 8px;
+    white-space: nowrap;
+    width: 80px;
+    display: inline-block;
+  }
+`;
