--- client/src/modules/Opportunities/components/forms/index.ts
@@ -1,5 +1,0 @@
-import ContactsForm from './ContactsForm';
-import UserDataForm from './GeneralInfoForm';
-import VisibleCoursesForm from './VisibleCoursesForm';
-
-export { ContactsForm, UserDataForm, VisibleCoursesForm };
